requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
Django==1.11.3
django-autoslug==1.9.3
django-crispy-forms==1.7.0
django-filter==1.0.4
django-pipeline==1.6.13
django-social-auth==0.7.28
django-webpack-loader==0.5.0
djangorestframework==3.6.4
djangorestframework-filters==0.10.1
futures==3.1.1
google-auth==1.3.0
httplib2==0.10.3
idna==2.6
oauth2==1.9.0.post1
olefile==0.44
Pillow==4.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
rsa==3.4.2
selenium==3.4.3
six==1.11.0
urllib3==1.22
